@@ -2,6 +2,7 @@
 
 #include "http_connection.h"
 #include <functional>
+#include <optional>
 
 namespace loki {
 using http_callback_t = std::function<void(sn_response_t)>;
@@ -40,8 +41,9 @@ class HttpsClientSession
     std::shared_ptr<request_t> req_;
     response_t res_;
 
-    // Snode's pub key (none if signature verification is not used / not a snode)
-    boost::optional<std::string> server_pub_key_b32z_;
+    // Snode's pub key (none if signature verification is not used / not a
+    // snode)
+    std::optional<std::string> server_pub_key_b32z_;
 
     bool used_callback_ = false;
 
@@ -53,7 +55,7 @@ class HttpsClientSession
 
     void
     trigger_callback(SNodeError error, std::shared_ptr<std::string>&& body,
-                     boost::optional<response_t> raw_response = boost::none);
+                     std::optional<response_t> raw_response = std::nullopt);
 
     void on_handshake(boost::system::error_code ec);
     bool verify_signature();
@@ -67,7 +69,7 @@ class HttpsClientSession
                        tcp::resolver::results_type resolve_results,
                        const std::shared_ptr<request_t>& req,
                        http_callback_t&& cb,
-                       boost::optional<const std::string&> sn_pubkey_b32z);
+                       std::optional<std::string> sn_pubkey_b32z);
 
     // initiate the client connection
     void start();
