@@ -32,15 +32,9 @@ class reach_record_t {
 };
 } // namespace detail
 
-enum class ReachType {
-  HTTP,
-  ZMQ
-};
+enum class ReachType { HTTP, ZMQ };
 
-enum class ReportType {
-  GOOD,
-  BAD
-};
+enum class ReportType { GOOD, BAD };
 
 class reachability_records_t {
 
@@ -53,7 +47,6 @@ class reachability_records_t {
     std::unordered_map<sn_pub_key_t, detail::reach_record_t> offline_nodes_;
 
   public:
-
     // The time we were last tested and reached by some other node over lmq
     time_point_t latest_incoming_lmq_;
     // The time we were last tested and reached by some other node over http
@@ -80,7 +73,7 @@ class reachability_records_t {
     void set_reported(const sn_pub_key_t& sn);
 
     // Retrun the least recently tested node
-    boost::optional<sn_pub_key_t> next_to_test();
+    std::optional<sn_pub_key_t> next_to_test();
 };
 
 } // namespace loki
