@@ -94,11 +94,11 @@ struct string_view {
     bool empty() { return it_end <= it; }
 };
 
-static boost::optional<std::string> deserialize_string(string_view& slice,
-                                                       size_t len) {
+static std::optional<std::string> deserialize_string(string_view& slice,
+                                                     size_t len) {
 
     if (slice.size() < len) {
-        return boost::none;
+        return std::nullopt;
     }
 
     const auto res = std::string(slice.it, slice.it + len);
@@ -107,10 +107,10 @@ static boost::optional<std::string> deserialize_string(string_view& slice,
     return res;
 }
 
-static boost::optional<std::string> deserialize_string(string_view& slice) {
+static std::optional<std::string> deserialize_string(string_view& slice) {
 
     if (slice.size() < sizeof(size_t))
-        return boost::none;
+        return std::nullopt;
 
     const auto len =
         deserialize_integer<size_t>(slice.it); // already increments `it`!
@@ -118,10 +118,10 @@ static boost::optional<std::string> deserialize_string(string_view& slice) {
     return deserialize_string(slice, len);
 }
 
-static boost::optional<uint64_t> deserialize_uint64(string_view& slice) {
+static std::optional<uint64_t> deserialize_uint64(string_view& slice) {
 
     if (slice.size() < sizeof(uint64_t))
-        return boost::none;
+        return std::nullopt;
 
     const auto res = deserialize_integer<uint64_t>(slice.it);
 
