@@ -22,17 +22,17 @@
 #include <algorithm>
 #include <chrono>
 #include <fstream>
+#include <string_view>
 
 #include <boost/bind.hpp>
 
 using json = nlohmann::json;
 using loki::storage::Item;
+using std::string_view;
 using namespace std::chrono_literals;
 
 namespace loki {
 using http_server::connection_t;
-
-using LockGuard = std::lock_guard<std::recursive_mutex>;
 
 constexpr std::array<std::chrono::seconds, 8> RETRY_INTERVALS = {
     1s, 5s, 10s, 20s, 40s, 80s, 160s, 320s};
@@ -47,10 +47,10 @@ static void make_sn_request(boost::asio::io_context& ioc, const sn_record_t& sn,
                        std::move(cb));
 }
 
-FailedRequestHandler::FailedRequestHandler(
-    boost::asio::io_context& ioc, const sn_record_t& sn,
-    std::shared_ptr<request_t> req,
-    boost::optional<std::function<void()>>&& give_up_cb)
+FailedRequestHandler::FailedRequestHandler(boost::asio::io_context& ioc,
+                                           const sn_record_t& sn,
+                                           std::shared_ptr<request_t> req,
+                                           std::function<void()> give_up_cb)
     : ioc_(ioc), retry_timer_(ioc), sn_(sn), request_(std::move(req)),
       give_up_callback_(std::move(give_up_cb)) {}
 
@@ -60,7 +60,7 @@ void FailedRequestHandler::retry(std::shared_ptr<FailedRequestHandler>&& self) {
     if (attempt_count_ > RETRY_INTERVALS.size()) {
         LOKI_LOG(debug, "Gave up after {} attempts", attempt_count_);
         if (give_up_callback_)
-            (*give_up_callback_)();
+            give_up_callback_();
         return;
     }
 
@@ -69,7 +69,7 @@ void FailedRequestHandler::retry(std::shared_ptr<FailedRequestHandler>&& self) {
              RETRY_INTERVALS[attempt_count_ - 1].count());
 
     retry_timer_.async_wait(
-        [self = std::move(self)](const boost::system::error_code& ec) mutable {
+        [self = std::move(self)](const boost::system::error_code&) mutable {
             /// Save some references before possibly moved out of `self`
             const auto& sn = self->sn_;
             auto& ioc = self->ioc_;
@@ -149,6 +149,7 @@ ServiceNode::ServiceNode(boost::asio::io_context& ioc,
                          boost::asio::io_context& worker_ioc, uint16_t port,
                          LokimqServer& lmq_server,
                          const lokid_key_pair_t& lokid_key_pair,
+                         const std::string& ed25519hex,
                          const std::string& db_location,
                          LokidClient& lokid_client, const bool force_start)
     : ioc_(ioc), worker_ioc_(worker_ioc),
@@ -172,7 +173,7 @@ ServiceNode::ServiceNode(boost::asio::io_context& ioc,
 
     // TODO: get rid of "unused" fields
     our_address_ = sn_record_t(port, lmq_server.port(), addr, pk_hex, "unused",
-                               "unused", "unused", "1.1.1.1");
+                               "unused", ed25519hex, "1.1.1.1");
 
     // TODO: fail hard if we can't encode our public key
     LOKI_LOG(info, "Read our snode address: {}", our_address_);
@@ -188,9 +189,9 @@ ServiceNode::ServiceNode(boost::asio::io_context& ioc,
     lokid_ping_timer_tick();
     cleanup_timer_tick();
 
-// #ifndef INTEGRATION_TEST
+#ifndef INTEGRATION_TEST
     ping_peers_tick();
-// #endif
+#endif
 
     worker_thread_ = boost::thread([this]() { worker_ioc_.run(); });
     boost::asio::post(worker_ioc_, [this]() {
@@ -302,7 +303,8 @@ parse_swarm_update(const std::shared_ptr<std::string>& response_body) {
                 swarm_map[swarm_id].push_back(sn);
 
                 if (pubkey_x25519_bin.size() == 32)
-                    bu.active_x25519_pubkeys.insert(std::move(pubkey_x25519_bin));
+                    bu.active_x25519_pubkeys.insert(
+                        std::move(pubkey_x25519_bin));
             }
         }
 
@@ -320,7 +322,7 @@ parse_swarm_update(const std::shared_ptr<std::string>& response_body) {
 
 void ServiceNode::bootstrap_data() {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     LOKI_LOG(trace, "Bootstrapping peer data");
 
@@ -400,9 +402,9 @@ void ServiceNode::bootstrap_data() {
     }
 }
 
-bool ServiceNode::snode_ready(boost::optional<std::string&> reason) {
+bool ServiceNode::snode_ready(std::string* reason) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     bool ready = true;
     std::string buf;
@@ -431,13 +433,23 @@ ServiceNode::~ServiceNode() {
     worker_thread_.join();
 };
 
-void ServiceNode::send_onion_to_sn(const sn_record_t& sn,
-                                   const std::string& payload,
-                                   const std::string& eph_key,
-                                   ss_client::Callback cb) const {
+void ServiceNode::send_onion_to_sn_v1(const sn_record_t& sn,
+                                      const std::string& payload,
+                                      const std::string& eph_key,
+                                      ss_client::Callback cb) const {
+
+    lmq_server_->request(sn.pubkey_x25519_bin(), "sn.onion_req", std::move(cb),
+                         lokimq::send_option::request_timeout{10s}, eph_key,
+                         payload);
+}
+
+void ServiceNode::send_onion_to_sn_v2(const sn_record_t& sn,
+                                      const std::string& payload,
+                                      const std::string& eph_key,
+                                      ss_client::Callback cb) const {
 
     lmq_server_->request(
-        sn.pubkey_x25519_bin(), "sn.onion_req", std::move(cb),
+        sn.pubkey_x25519_bin(), "sn.onion_req_v2", std::move(cb),
         lokimq::send_option::request_timeout{10s}, eph_key, payload);
 }
 
@@ -446,14 +458,14 @@ void ServiceNode::send_to_sn(const sn_record_t& sn, ss_client::ReqMethod method,
                              ss_client::Request req,
                              ss_client::Callback cb) const {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     switch (method) {
     case ss_client::ReqMethod::DATA: {
         LOKI_LOG(debug, "Sending sn.data request to {}",
                  util::as_hex(sn.pubkey_x25519_bin()));
-        lmq_server_->request(sn.pubkey_x25519_bin(), "sn.data",
-                                   std::move(cb), req.body);
+        lmq_server_->request(sn.pubkey_x25519_bin(), "sn.data", std::move(cb),
+                             req.body);
         break;
     }
     case ss_client::ReqMethod::PROXY_EXIT: {
@@ -465,8 +477,7 @@ void ServiceNode::send_to_sn(const sn_record_t& sn, ss_client::ReqMethod method,
             LOKI_LOG(debug, "Sending sn.proxy_exit request to {}",
                      util::as_hex(sn.pubkey_x25519_bin()));
             lmq_server_->request(sn.pubkey_x25519_bin(), "sn.proxy_exit",
-                                       std::move(cb), client_key->second,
-                                       req.body);
+                                 std::move(cb), client_key->second, req.body);
         } else {
             LOKI_LOG(debug, "Developer error: no {} passed in headers",
                      LOKI_SENDER_KEY_HEADER);
@@ -509,7 +520,7 @@ void ServiceNode::record_onion_request() { all_stats_.bump_onion_requests(); }
 /// do this asynchronously on a different thread? (on the same thread?)
 bool ServiceNode::process_store(const message_t& msg) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     /// only accept a message if we are in a swarm
     if (!swarm_) {
@@ -532,7 +543,7 @@ bool ServiceNode::process_store(const message_t& msg) {
 
 void ServiceNode::save_if_new(const message_t& msg) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     if (db_->store(msg.hash, msg.pub_key, msg.data, msg.ttl, msg.timestamp,
                    msg.nonce)) {
@@ -542,7 +553,7 @@ void ServiceNode::save_if_new(const message_t& msg) {
 
 void ServiceNode::save_bulk(const std::vector<Item>& items) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     if (!db_->bulk_store(items)) {
         LOKI_LOG(error, "failed to save batch to the database");
@@ -555,7 +566,7 @@ void ServiceNode::save_bulk(const std::vector<Item>& items) {
 void ServiceNode::on_bootstrap_update(block_update_t&& bu) {
 
     // Used in a callback to needs a mutex even if it is private
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     swarm_->apply_swarm_changes(bu.swarms);
     target_height_ = std::max(target_height_, bu.height);
@@ -578,34 +589,35 @@ OStream& operator<<(OStream& os, const SnodeStatus& status) {
     }
 }
 
-static SnodeStatus derive_snode_status(const block_update_t& bu, const sn_record_t& our_address) {
+static SnodeStatus derive_snode_status(const block_update_t& bu,
+                                       const sn_record_t& our_address) {
 
     // TODO: try not to do this again in `derive_swarm_events`
-    const auto our_swarm_it = std::find_if(
-        bu.swarms.begin(), bu.swarms.end(), [&our_address](const SwarmInfo& swarm_info) {
-            const auto& snodes = swarm_info.snodes;
-            return std::find(snodes.begin(), snodes.end(), our_address) !=
-                   snodes.end();
-        });
+    const auto our_swarm_it =
+        std::find_if(bu.swarms.begin(), bu.swarms.end(),
+                     [&our_address](const SwarmInfo& swarm_info) {
+                         const auto& snodes = swarm_info.snodes;
+                         return std::find(snodes.begin(), snodes.end(),
+                                          our_address) != snodes.end();
+                     });
 
     if (our_swarm_it != bu.swarms.end()) {
         return SnodeStatus::ACTIVE;
     }
 
-    if (std::find(bu.decommissioned_nodes.begin(), bu.decommissioned_nodes.end(), our_address) != bu.decommissioned_nodes.end()) {
+    if (std::find(bu.decommissioned_nodes.begin(),
+                  bu.decommissioned_nodes.end(),
+                  our_address) != bu.decommissioned_nodes.end()) {
         return SnodeStatus::DECOMMISSIONED;
     }
 
     return SnodeStatus::UNSTAKED;
-
 }
-
-
 
 void ServiceNode::on_swarm_update(block_update_t&& bu) {
 
     // Used in a callback to needs a mutex even if it is private
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     if (this->hardfork_ != bu.hardfork) {
         LOKI_LOG(debug, "New hardfork: {}", bu.hardfork);
@@ -666,7 +678,7 @@ void ServiceNode::on_swarm_update(block_update_t&& bu) {
     swarm_->set_swarm_id(events.our_swarm_id);
 
     std::string reason;
-    if (!this->snode_ready(boost::optional<std::string&>(reason))) {
+    if (!this->snode_ready(&reason)) {
         LOKI_LOG(warn, "Storage server is still not ready: {}", reason);
         swarm_->update_state(bu.swarms, bu.decommissioned_nodes, events, false);
         return;
@@ -708,7 +720,7 @@ void ServiceNode::on_swarm_update(block_update_t&& bu) {
 
 void ServiceNode::relay_buffered_messages() {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     // Should we wait for the response first?
     relay_timer_.expires_after(RELAY_INTERVAL);
@@ -747,7 +759,7 @@ void ServiceNode::pow_difficulty_timer_tick(const pow_dns_callback_t cb) {
 
 void ServiceNode::swarm_timer_tick() {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     LOKI_LOG(trace, "Swarm timer tick");
 
@@ -811,7 +823,7 @@ void ServiceNode::swarm_timer_tick() {
 
 void ServiceNode::cleanup_timer_tick() {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     all_stats_.cleanup();
 
@@ -823,25 +835,25 @@ void ServiceNode::cleanup_timer_tick() {
 void ServiceNode::update_last_ping(ReachType type) {
 
     switch (type) {
-        case ReachType::HTTP: {
-            reach_records_.latest_incoming_http_ = std::chrono::steady_clock::now();
-            break;
-        }
-        case ReachType::ZMQ: {
-            reach_records_.latest_incoming_lmq_ = std::chrono::steady_clock::now();
-            break;
-        }
-        default:
-            LOKI_LOG(error, "Connection type not supported");
-            assert(false);
-            break;
+    case ReachType::HTTP: {
+        reach_records_.latest_incoming_http_ = std::chrono::steady_clock::now();
+        break;
+    }
+    case ReachType::ZMQ: {
+        reach_records_.latest_incoming_lmq_ = std::chrono::steady_clock::now();
+        break;
+    }
+    default:
+        LOKI_LOG(error, "Connection type not supported");
+        assert(false);
+        break;
     }
 }
 
 void ServiceNode::ping_peers_tick() {
 
     // Used as a callback to needs a mutex even if it is private
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     this->peer_ping_timer_.expires_after(PING_PEERS_INTERVAL);
     this->peer_ping_timer_.async_wait(
@@ -849,7 +861,8 @@ void ServiceNode::ping_peers_tick() {
 
     // TODO: Don't do anything until we are fully funded
 
-    if (this->status_ == SnodeStatus::UNSTAKED || this->status_ == SnodeStatus::UNKNOWN) {
+    if (this->status_ == SnodeStatus::UNSTAKED ||
+        this->status_ == SnodeStatus::UNKNOWN) {
         LOKI_LOG(debug, "Skipping this round of peer testing (unstaked)");
         return;
     }
@@ -887,7 +900,7 @@ void ServiceNode::ping_peers_tick() {
     const auto offline_node = reach_records_.next_to_test();
 
     if (offline_node) {
-        const boost::optional<sn_record_t> sn =
+        const std::optional<sn_record_t> sn =
             swarm_->get_node_by_pk(*offline_node);
         LOKI_LOG(debug, "No offline nodes to test for reachability yet");
         if (sn) {
@@ -903,7 +916,7 @@ void ServiceNode::ping_peers_tick() {
 
 void ServiceNode::sign_request(std::shared_ptr<request_t>& req) const {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     // TODO: investigate why we are not signing headers
     const auto hash = hash_data(req->body());
@@ -913,7 +926,7 @@ void ServiceNode::sign_request(std::shared_ptr<request_t>& req) const {
 
 void ServiceNode::test_reachability(const sn_record_t& sn) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     LOKI_LOG(debug, "Testing node for reachability over HTTP: {}", sn);
 
@@ -934,22 +947,22 @@ void ServiceNode::test_reachability(const sn_record_t& sn) {
     LOKI_LOG(debug, "Testing node for reachability over LMQ: {}", sn);
 
     // test lmq port:
-    lmq_server_->request(
-        sn.pubkey_x25519_bin(), "sn.onion_req",
-        [this, sn](bool success, const auto&) {
-            LOKI_LOG(debug, "Got success={} testing response from {}", success,
-                     sn.pubkey_x25519_hex());
-            this->process_reach_test_result(sn.pub_key_base32z(),
-                                            ReachType::ZMQ, success);
-        },
-        "ping",
-        // Only use an existing (or new) outgoing connection:
-        lokimq::send_option::outgoing{});
+    lmq_server_->request(sn.pubkey_x25519_bin(), "sn.onion_req",
+                         [this, sn](bool success, const auto&) {
+                             LOKI_LOG(debug,
+                                      "Got success={} testing response from {}",
+                                      success, sn.pubkey_x25519_hex());
+                             this->process_reach_test_result(
+                                 sn.pub_key_base32z(), ReachType::ZMQ, success);
+                         },
+                         "ping",
+                         // Only use an existing (or new) outgoing connection:
+                         lokimq::send_option::outgoing{});
 }
 
 void ServiceNode::lokid_ping_timer_tick() {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     /// TODO: Note that this is not actually an SN response! (but Lokid)
     auto cb = [](const sn_response_t&& res) {
@@ -1012,7 +1025,7 @@ void ServiceNode::perform_blockchain_test(
     bc_test_params_t test_params,
     std::function<void(blockchain_test_answer_t)>&& cb) const {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     LOKI_LOG(debug, "Delegating blockchain test to Lokid");
 
@@ -1075,7 +1088,7 @@ void ServiceNode::process_storage_test_response(const sn_record_t& testee,
                                                 uint64_t test_height,
                                                 sn_response_t&& res) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     if (res.error_code != SNodeError::NO_ERROR) {
         // TODO: retry here, otherwise tests sometimes fail (when SN not
@@ -1140,7 +1153,7 @@ void ServiceNode::send_storage_test_req(const sn_record_t& testee,
                                         const Item& item) {
 
     // Used as a callback to needs a mutex even if it is private
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     nlohmann::json json_body;
 
@@ -1165,7 +1178,7 @@ void ServiceNode::send_blockchain_test_req(const sn_record_t& testee,
                                            blockchain_test_answer_t answer) {
 
     // Used as a callback to needs a mutex even if it is private
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     nlohmann::json json_body;
 
@@ -1186,7 +1199,7 @@ void ServiceNode::send_blockchain_test_req(const sn_record_t& testee,
 void ServiceNode::report_node_reachability(const sn_pub_key_t& sn_pk,
                                            bool reachable) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     const auto sn = swarm_->get_node_by_pk(sn_pk);
 
@@ -1204,7 +1217,7 @@ void ServiceNode::report_node_reachability(const sn_pub_key_t& sn_pk,
     /// updated to "true".
 
     auto cb = [this, sn_pk, reachable](const sn_response_t&& res) {
-        LockGuard guard(this->sn_mutex_);
+        std::lock_guard guard(this->sn_mutex_);
 
         if (res.error_code != SNodeError::NO_ERROR) {
             LOKI_LOG(warn, "Could not report node status");
@@ -1254,7 +1267,7 @@ void ServiceNode::report_node_reachability(const sn_pub_key_t& sn_pk,
 void ServiceNode::process_reach_test_result(const sn_pub_key_t& pk,
                                             ReachType type, bool success) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     if (success) {
 
@@ -1285,7 +1298,7 @@ void ServiceNode::process_blockchain_test_response(
     sn_response_t&& res, blockchain_test_answer_t our_answer,
     sn_record_t testee, uint64_t bc_height) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     LOKI_LOG(debug,
              "Processing blockchain test response from: {} at height: {}",
@@ -1324,7 +1337,7 @@ void ServiceNode::process_blockchain_test_response(
 bool ServiceNode::derive_tester_testee(uint64_t blk_height, sn_record_t& tester,
                                        sn_record_t& testee) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     std::vector<sn_record_t> members = swarm_->other_nodes();
     members.push_back(our_address_);
@@ -1389,7 +1402,7 @@ MessageTestStatus ServiceNode::process_storage_test_req(
     uint64_t blk_height, const std::string& tester_pk,
     const std::string& msg_hash, std::string& answer) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     // 1. Check height, retry if we are behind
     std::string block_hash;
@@ -1460,7 +1473,7 @@ bool ServiceNode::select_random_message(Item& item) {
 
 void ServiceNode::initiate_peer_test() {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     // 1. Select the tester/testee pair
     sn_record_t tester, testee;
@@ -1569,7 +1582,7 @@ std::string vec_to_string(const std::vector<T>& vec) {
 void ServiceNode::bootstrap_swarms(
     const std::vector<swarm_id_t>& swarms) const {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     if (swarms.empty()) {
         LOKI_LOG(info, "Bootstrapping all swarms");
@@ -1676,7 +1689,7 @@ bool ServiceNode::retrieve(const std::string& pubKey,
                            const std::string& last_hash,
                            std::vector<Item>& items) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     all_stats_.bump_retrieve_requests();
 
@@ -1687,7 +1700,7 @@ bool ServiceNode::retrieve(const std::string& pubKey,
 void ServiceNode::set_difficulty_history(
     const std::vector<pow_difficulty_t>& new_history) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     pow_history_ = new_history;
     for (const auto& difficulty : pow_history_) {
@@ -1717,6 +1730,9 @@ static nlohmann::json to_json(const all_stats_t& stats) {
     json["previous_period_retrieve_requests"] =
         stats.get_previous_period_retrieve_requests();
 
+    json["previous_period_onion_requests"] =
+        stats.get_previous_period_onion_requests();
+
     json["reset_time"] = std::chrono::duration_cast<std::chrono::seconds>(
                              stats.get_reset_time().time_since_epoch())
                              .count();
@@ -1736,9 +1752,19 @@ static nlohmann::json to_json(const all_stats_t& stats) {
     return json;
 }
 
+std::string ServiceNode::get_stats_for_session_client() const {
+
+    nlohmann::json res;
+    res["version"] = STORAGE_SERVER_VERSION_STRING;
+
+    constexpr bool PRETTY = true;
+    constexpr int indent = PRETTY ? 4 : 0;
+    return res.dump(indent);
+}
+
 std::string ServiceNode::get_stats() const {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     auto val = to_json(all_stats_);
 
@@ -1754,7 +1780,6 @@ std::string ServiceNode::get_stats() const {
     val["connections_in"] = get_net_stats().connections_in.load();
     val["http_connections_out"] = get_net_stats().http_connections_out.load();
     val["https_connections_out"] = get_net_stats().https_connections_out.load();
-    // val["open_socket_count"] = get_net_stats().open_fds.size();
 
     /// we want pretty (indented) json, but might change that in the future
     constexpr bool PRETTY = true;
@@ -1768,7 +1793,7 @@ std::string ServiceNode::get_status_line() const {
     // status message has to be fairly short: has to fit on one line, and if
     // it's too long systemd just truncates it when displaying it.
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     std::ostringstream s;
     s << 'v' << STORAGE_SERVER_VERSION_STRING;
@@ -1801,14 +1826,14 @@ std::string ServiceNode::get_status_line() const {
 
 int ServiceNode::get_curr_pow_difficulty() const {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     return curr_pow_difficulty_.difficulty;
 }
 
 bool ServiceNode::get_all_messages(std::vector<Item>& all_entries) const {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     LOKI_LOG(trace, "Get all messages");
 
@@ -1817,7 +1842,7 @@ bool ServiceNode::get_all_messages(std::vector<Item>& all_entries) const {
 
 void ServiceNode::process_push_batch(const std::string& blob) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     if (blob.empty())
         return;
@@ -1861,7 +1886,7 @@ void ServiceNode::process_push_batch(const std::string& blob) {
 
 bool ServiceNode::is_pubkey_for_us(const user_pubkey_t& pk) const {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     if (!swarm_) {
         LOKI_LOG(error, "Swarm data missing");
@@ -1873,7 +1898,7 @@ bool ServiceNode::is_pubkey_for_us(const user_pubkey_t& pk) const {
 std::vector<sn_record_t>
 ServiceNode::get_snodes_by_pk(const user_pubkey_t& pk) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     if (!swarm_) {
         LOKI_LOG(error, "Swarm data missing");
@@ -1899,7 +1924,7 @@ ServiceNode::get_snodes_by_pk(const user_pubkey_t& pk) {
 
 bool ServiceNode::is_snode_address_known(const std::string& sn_address) {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     // TODO: need more robust handling of uninitialized swarm_
     if (!swarm_) {
@@ -1910,28 +1935,28 @@ bool ServiceNode::is_snode_address_known(const std::string& sn_address) {
     return swarm_->is_fully_funded_node(sn_address);
 }
 
-boost::optional<sn_record_t>
+std::optional<sn_record_t>
 ServiceNode::find_node_by_x25519_bin(const sn_pub_key_t& pk) const {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     if (swarm_) {
         return swarm_->find_node_by_x25519_bin(pk);
     }
 
-    return boost::none;
+    return std::nullopt;
 }
 
-boost::optional<sn_record_t>
+std::optional<sn_record_t>
 ServiceNode::find_node_by_ed25519_pk(const std::string& pk) const {
 
-    LockGuard guard(sn_mutex_);
+    std::lock_guard guard(sn_mutex_);
 
     if (swarm_) {
         return swarm_->find_node_by_ed25519_pk(pk);
     }
 
-    return boost::none;
+    return std::nullopt;
 }
 
 } // namespace loki
