@@ -16,9 +16,10 @@ namespace loki {
 
 private_key_t lokidKeyFromHex(const std::string& private_key_hex) {
     if (private_key_hex.size() != KEY_LENGTH * 2)
-        throw std::runtime_error(
-                "Lokid key data is invalid: expected " + std::to_string(KEY_LENGTH) + " bytes not " +
-                std::to_string(private_key_hex.size()) + " bytes");
+        throw std::runtime_error("Lokid key data is invalid: expected " +
+                                 std::to_string(KEY_LENGTH) + " bytes not " +
+                                 std::to_string(private_key_hex.size()) +
+                                 " bytes");
 
     const auto bytes = util::hex_to_bytes(private_key_hex);
     private_key_t private_key;
@@ -45,11 +46,11 @@ private_key_ed25519_t::from_hex(const std::string& sc_hex) {
 
 public_key_t derive_pubkey_legacy(const private_key_t& private_key) {
     public_key_t publicKey;
-    crypto_scalarmult_ed25519_base_noclamp(publicKey.data(), private_key.data());
+    crypto_scalarmult_ed25519_base_noclamp(publicKey.data(),
+                                           private_key.data());
 
     return publicKey;
 }
-
 
 public_key_t derive_pubkey_x25519(const private_key_t& seckey) {
 
