@@ -118,8 +118,7 @@ bool check_signature(const signature& sig, const hash& prefix_hash,
 bool check_signature(const std::string& signature, const hash& hash,
                      const std::string& public_key_b32z) {
     // convert signature
-    const std::string raw_signature =
-        util::base64_decode(signature);
+    const std::string raw_signature = util::base64_decode(signature);
     struct signature sig;
     std::copy_n(raw_signature.begin(), sig.c.size(), sig.c.begin());
     std::copy_n(raw_signature.begin() + sig.c.size(), sig.r.size(),
