@@ -18,8 +18,10 @@ struct command_line_options {
     std::string log_level = "info";
     std::string data_dir;
     std::string lokid_key; // test only (but needed for backwards compatibility)
-    std::string lokid_x25519_key; // test only
+    std::string lokid_x25519_key;  // test only
     std::string lokid_ed25519_key; // test only
+    // x25519 key that will be given access to get_stats lmq endpoint
+    std::vector<std::string> stats_access_keys;
 };
 
 class command_line_parser {
