@@ -3,6 +3,7 @@
 #include "utils.hpp"
 
 #include "sqlite3.h"
+#include <cstdlib>
 #include <exception>
 
 namespace loki {
@@ -68,6 +69,78 @@ sqlite3_stmt* Database::prepare_statement(const std::string& query) {
     return stmt;
 }
 
+constexpr int64_t DB_PAGE_SIZE = 4096;
+constexpr int64_t DB_SIZE_LIMIT = int64_t(3584) * 1024 * 1024; // 3.5 GB
+constexpr int64_t DB_PAGE_LIMIT = DB_SIZE_LIMIT / DB_PAGE_SIZE;
+
+static void set_page_count(sqlite3* db) {
+
+    char* errMsg = nullptr;
+
+    auto cb = [](void* a_param, int argc, char** argv, char** column) -> int {
+        if (argc == 0) {
+            LOKI_LOG(error, "Failed to set the page count limit");
+            return 0;
+        }
+
+        int res = strtol(argv[0], NULL, 10);
+
+        if (res == 0) {
+            LOKI_LOG(error, "Failed to convert page limit ({}) to a number",
+                     argv[0]);
+            return 0;
+        }
+
+        LOKI_LOG(info, "DB page limit is set to: {}", res);
+
+        return 0;
+    };
+
+    int rc = sqlite3_exec(
+        db, fmt::format("PRAGMA MAX_PAGE_COUNT = {};", DB_PAGE_LIMIT).c_str(),
+        cb, nullptr, &errMsg);
+
+    if (rc) {
+        if (errMsg) {
+            LOKI_LOG(error, "Query error: {}", errMsg);
+        }
+    }
+}
+
+static void check_page_size(sqlite3* db) {
+
+    char* errMsg = nullptr;
+
+    auto cb = [](void* a_param, int argc, char** argv, char** column) -> int {
+        if (argc == 0) {
+            LOKI_LOG(error, "Could not get DB page size");
+        }
+
+        int res = strtol(argv[0], NULL, 10);
+
+        if (res == 0) {
+            LOKI_LOG(error, "Failed to convert page size ({}) to a number",
+                     argv[0]);
+            return 0;
+        }
+
+        if (res != DB_PAGE_SIZE) {
+            LOKI_LOG(warn, "Unexpected DB page size: {}", res);
+        } else {
+            LOKI_LOG(info, "DB page size: {}", res);
+        }
+
+        return 0;
+    };
+
+    int rc = sqlite3_exec(db, "PRAGMA page_size;", cb, nullptr, &errMsg);
+    if (rc) {
+        if (errMsg) {
+            LOKI_LOG(error, "Query error: {}", errMsg);
+        }
+    }
+}
+
 void Database::open_and_prepare(const std::string& db_path) {
     const std::string file_path = db_path + "/storage.db";
     int rc = sqlite3_open_v2(file_path.c_str(), &db,
@@ -81,6 +154,9 @@ void Database::open_and_prepare(const std::string& db_path) {
         // throw?
         return;
     }
+
+    check_page_size(db);
+    set_page_count(db);
 
     const char* create_table_query =
         "CREATE TABLE IF NOT EXISTS `Data`("
@@ -293,6 +369,11 @@ bool Database::store(const std::string& hash, const std::string& pubKey,
     sqlite3_bind_blob(stmt, 6, nonce.data(), nonce.size(), SQLITE_STATIC);
     sqlite3_bind_blob(stmt, 7, bytes.data(), bytes.size(), SQLITE_STATIC);
 
+    // keep track of db full errorss so we don't print them on every store
+    static int db_full_counter = 0;
+    // print the error once so many errors
+    constexpr int DB_FULL_FREQUENCY = 100;
+
     bool result = false;
     int rc;
     while (true) {
@@ -304,6 +385,12 @@ bool Database::store(const std::string& hash, const std::string& pubKey,
         } else if (rc == SQLITE_DONE) {
             result = true;
             break;
+        } else if (rc == SQLITE_FULL) {
+            if (db_full_counter % DB_FULL_FREQUENCY == 0) {
+                LOKI_LOG(error, "Failed to store message: database is full");
+                ++db_full_counter;
+            }
+            break;
         } else {
             LOKI_LOG(critical, "Could not execute `store` db statement, ec: {}",
                      rc);
@@ -312,7 +399,7 @@ bool Database::store(const std::string& hash, const std::string& pubKey,
     }
 
     rc = sqlite3_reset(stmt);
-    if (rc != SQLITE_OK && rc != SQLITE_CONSTRAINT) {
+    if (rc != SQLITE_OK && rc != SQLITE_CONSTRAINT && rc != SQLITE_FULL) {
         LOKI_LOG(critical, "sqlite reset error: [{}], {}", rc,
                  sqlite3_errmsg(db));
     }
